Casted unsigned-to-signed comparisons.
svn path=/trunk/KDE/kdebase/apps/; revision=939420

+++ kfind/kquery.cpp
@@ -208,9 +208,9 @@ void KQuery::processQuery( const KFileItem &file)
 
     // make sure it's in the correct date range
     // what about 0 times?
-    if ( m_timeFrom && m_timeFrom > file.time(KFileItem::ModificationTime).toTime_t() )
+    if ( m_timeFrom && ((uint) m_timeFrom) > file.time(KFileItem::ModificationTime).toTime_t() )
       return;
-    if ( m_timeTo && m_timeTo < file.time(KFileItem::ModificationTime).toTime_t() )
+    if ( m_timeTo && ((uint) m_timeTo) < file.time(KFileItem::ModificationTime).toTime_t() )
       return;
 
     // username / group match
